requirements.txt
openai
python-dotenv
groq

-e .